+++ tests/test_arcs.py
@@ -926,11 +926,6 @@ class ExceptionArcTest(CoverageTest):
             arcz=".1 12 23 35 56 61 17 7.",
         )
 
-    @pytest.mark.xfail(
-        env.PYPY and env.PYVERSION >= (3, 9),
-        reason="avoid a PyPy bug: 3662"
-        # https://foss.heptapod.net/pypy/pypy/-/issues/3662
-    )
     def test_bug_212(self):
         # "except Exception as e" is crucial here.
         # Bug 212 said that the "if exc" line was incorrectly marked as only
@@ -1653,11 +1648,6 @@ class MiscArcTest(CoverageTest):
 class DecoratorArcTest(CoverageTest):
     """Tests of arcs with decorators."""
 
-    @pytest.mark.xfail(
-        env.PYPY and env.PYVERSION >= (3, 9),
-        reason="avoid a PyPy bug: 3666"
-        # https://foss.heptapod.net/pypy/pypy/-/issues/3666
-    )
     def test_function_decorator(self):
         arcz = (
             ".1 16 67 7A AE EF F. "     # main line
@@ -1686,11 +1676,6 @@ class DecoratorArcTest(CoverageTest):
             arcz=arcz,
         )
 
-    @pytest.mark.xfail(
-        env.PYPY and env.PYVERSION >= (3, 9),
-        reason="avoid a PyPy bug: 3666"
-        # https://foss.heptapod.net/pypy/pypy/-/issues/3666
-    )
     def test_class_decorator(self):
         arcz = (
             ".1 16 67 6D 7A AE E. "     # main line