+++ CHANGES.rst
@@ -17,8 +17,10 @@ development at the same time, such as 4.5.x and 5.0.
     ..  Version 9.8.1 — 2027-07-27
     ..  --------------------------
 
-Unreleased
-----------
+.. _changes_632:
+
+Version 6.3.2 — 2022-02-20
+--------------------------
 
 - Fix: adapt to pypy3.9's decorator tracing behavior.  It now traces function
   decorators like CPython 3.8: both the @-line and the def-line are traced.